requirements.txt
python-dotenv
requests
pydantic
langchain-ollama
